fix: Move pkg/wallet/wallets to internal/wallet to resolve Go convention violation
This commit fixes the review comment
https://github.com/hiromaily/go-crypto-wallet/pull/78#discussion_r2646445034

Changes:
- Moved pkg/wallet/wallets/ to internal/wallet/
- Updated all import paths from pkg/wallet/wallets to internal/wallet
- Affected 12 wallet implementation files and 20 files that import them

The wallet implementations (btcwallet, ethwallet, xrpwallet) now import
from internal/application/usecase and internal/domain packages, so they
must be in the internal/ directory to comply with Go conventions where
pkg/ should not depend on internal/.

This aligns with Clean Architecture by keeping application services
that orchestrate use cases within the internal/ directory.

Related to #77 (Issue #77 - Directory Structure Refactoring)
