+++ cmd/watch/main.go
@@ -13,9 +13,9 @@ import (
 	domainCoin "github.com/hiromaily/go-crypto-wallet/internal/domain/coin"
 	domainWallet "github.com/hiromaily/go-crypto-wallet/internal/domain/wallet"
 	wcmd "github.com/hiromaily/go-crypto-wallet/internal/interface-adapters/cli/watch"
+	wallets "github.com/hiromaily/go-crypto-wallet/internal/wallet"
 	"github.com/hiromaily/go-crypto-wallet/pkg/account"
 	"github.com/hiromaily/go-crypto-wallet/pkg/config"
-	"github.com/hiromaily/go-crypto-wallet/pkg/wallet/wallets"
 )
 
 // watch as watch only wallet
