+++ internal/interface-adapters/cli/sign/export/export.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/hiromaily/go-crypto-wallet/internal/di"
-	"github.com/hiromaily/go-crypto-wallet/pkg/wallet/wallets"
+	wallets "github.com/hiromaily/go-crypto-wallet/internal/wallet"
 )
 
 // AddCommands adds all export subcommands
