+++ internal/wallet/btcwallet/keygen.go
