+++ internal/wallet/btcwallet/sign.go
