+++ internal/wallet/signer.go
